Add authorization check (permissions), fix profile/user views

+++ products/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
 from django.conf import settings
 from products.models import ProductModel, ImageModel
+from django.contrib.auth import get_user_model
 from products.serializers import (
     ProductModelSerializer,
     ImageModelSerializer,
@@ -9,18 +10,19 @@ from products.serializers import (
     ProductModelUpdateSerializer,
 )
 from rest_framework.views import APIView
-from rest_framework.response import Response
 from rest_framework import generics, filters
 from rest_framework.permissions import IsAuthenticated
+from .permissions import IsOwnerOrReadOnly
+
 import stripe
 
+User = get_user_model()
 stripe.api_key = settings.STRIPE_SECRET_KEY
 
 # Create your views here.
 class LatestView(generics.ListAPIView):
     queryset = ProductModel.objects.all()
     serializer_class = ProductModelSerializer
-    permission_classes = [IsAuthenticated]
 
 
 class Details1View(generics.ListAPIView):
@@ -59,42 +61,66 @@ class SearchView(generics.ListAPIView):
 
 class ProfileView(generics.ListAPIView):
     serializer_class = ProductModelSerializer
+    lookup_field = "profile_id"
+    permission_classes = [IsAuthenticated, IsOwnerOrReadOnly]
 
     def get_queryset(self):
-        user_id = self.kwargs["user_id"]
+        profile_id = self.kwargs["profile_id"]
+        products = ProductModel.objects.filter(owner=profile_id)
+        # for product in products:
+        #     self.check_object_permissions(self.request, product)
+        return products
+
+
+class UserView(generics.ListAPIView):
+    serializer_class = ProductModelSerializer
+    lookup_field = "username"
+
+    def get_queryset(self):
+        username = self.kwargs["username"]
+        user_id = User.objects.get(username=username)
         products = ProductModel.objects.filter(owner=user_id)
         return products
 
 
 class ProductModelCreateView(generics.CreateAPIView):
     serializer_class = ProductModelCreateSerializer
+    permission_classes = [IsAuthenticated, IsOwnerOrReadOnly]
 
 
 class ImageModelCreateView(generics.CreateAPIView):
     serializer_class = ImageModelCreateSerializer
+    permission_classes = [IsAuthenticated, IsOwnerOrReadOnly]
 
 
 class ProductModelUpdateView(generics.UpdateAPIView):
     queryset = ProductModel.objects.all()
     serializer_class = ProductModelUpdateSerializer
     lookup_field = "stripe_product_id"
+    permission_classes = [IsAuthenticated, IsOwnerOrReadOnly]
 
 
 class ProductModelDeleteView(generics.DestroyAPIView):
+    queryset = ProductModel.objects.all()
     lookup_field = "stripe_product_id"
+    permission_classes = [IsAuthenticated, IsOwnerOrReadOnly]
 
     def get_queryset(self):
         stripe_product_id = self.kwargs["stripe_product_id"]
         stripe.Product.modify(stripe_product_id, active="false")
-        product = ProductModel.objects.filter(stripe_product_id=stripe_product_id)
+        product = ProductModel.objects.get(stripe_product_id=stripe_product_id)
+        self.check_object_permissions(self.request, product)
         product.delete()
         images = ImageModel.objects.filter(stripe_product_id=stripe_product_id)
+        for image in images:
+            self.check_object_permissions(self.request, image)
         images.delete()
 
 
 class ImageModelDeleteView(generics.DestroyAPIView):
     queryset = ImageModel.objects.all()
     lookup_field = "id"
+    permission_classes = [IsAuthenticated, IsOwnerOrReadOnly]
 
 
 def success(request):
